Поменял местами Воду и Огонь
